pyproject.toml
[tool.poetry]
name = "chickbwk"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
"discord.py" = "^1.3.4"
aiohttp = "^3.6.2"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.3"
black = {version = "^19.10b0", allow-prereleases = true}

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
